requirements.txt
bokeh==2.4.3
pandas==1.5.3
Pillow==9.4.0
streamlit==1.23.1
